+++ Block Management/SCBlockEntry.h
@@ -12,11 +12,11 @@ NS_ASSUME_NONNULL_BEGIN
 @interface SCBlockEntry : NSObject
 
 @property (nonatomic) NSString* hostname;
-@property (nonatomic) int port;
-@property (nonatomic) int maskLen;
+@property (nonatomic) NSInteger port;
+@property (nonatomic) NSInteger maskLen;
 
 + (instancetype)entryWithHostname:(NSString*)hostname;
-+ (instancetype)entryWithHostname:(NSString*)hostname port:(int)port maskLen:(int)maskLen;
++ (instancetype)entryWithHostname:(NSString*)hostname port:(NSInteger)port maskLen:(NSInteger)maskLen;
 + (instancetype)entryFromString:(NSString*)domainString;
 
 - (BOOL)isEqualToEntry:(SCBlockEntry*)otherEntry;
